@@ -3,6 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-regular-svg-icons": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "lodash": "^4.17.21",
     "luxon": "^2.3.0",
     "react": "^18.2.0",
@@ -37,17 +41,17 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "@types/gapi": "^0.0.41",
-    "@types/gapi.auth2": "^0.0.55",
+    "@types/gapi": "^0.0.43",
     "@types/gapi.calendar": "^3.0.6",
     "@types/google.accounts": "^0.0.4",
     "@types/jest": "^27.4.0",
+    "@types/lodash": "^4.14.191",
     "@types/luxon": "^2.0.9",
     "@types/node": "^16.11.17",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.10",
     "react-scripts": "^5.0.0",
-    "typescript": "^4.5.4",
+    "typescript": "^4.9.4",
     "web-vitals": "^2.1.2"
   }
 }
