@@ -2,13 +2,11 @@ import { PROPS_FOR_STEP } from "../AppWorkflowStep";
 
 export function NavSidebar() {
     return <div>
-        <div className="text-black-50 fst-italic pb-1">Navigate to...</div>
+        <div className="text-body text-opacity-50 fst-italic pb-1">Navigate to...</div>
         <div className="list-group list-group-flush font-heading">
-            <a href="#top" className="list-group-item list-group-item-action">Top</a>
-            {Object.values(PROPS_FOR_STEP).map( ({ id, title }) =>
-                <a href={"#" + id} key={id} className="list-group-item list-group-item-action">{title}</a>
+            {Object.values(PROPS_FOR_STEP).map( ({ id, heading }) =>
+                <a href={"#" + id} key={id} className="list-group-item list-group-item-action">{heading}</a>
             )}
-            <a href="#about" className="list-group-item list-group-item-action">About this site</a>
         </div>
     </div>;
 }
